test: move tool provider controller coverage to unit tests
Reclassify unit-level coverage from tests/test_containers_integration_tests/controllers/console/workspace/test_tool_provider.py into the unit test suite, or remove it when equivalent unit coverage already exists. Preserve real container-backed coverage and use SQLite sessions for database-dependent unit paths.
